feat: migrate company profile storage from IndexedDB to Supabase
